Finalize MAIN source ownership relocation
Move the remaining MAIN-only Lit style foundation beneath the MAIN runtime, update all consumers, and align linting and repository guidance with the relocated component ownership.

Keep repository-level shared helpers intact for the dedicated #97 follow-up.
