+++ src/content/main/features/action-recorder/action-recorder-dialog.js
@@ -2,7 +2,7 @@ import { LitElement, html } from 'lit';
 import {
     componentFoundationStyles,
     dialogFoundationStyles
-} from '../../../../components/styles/foundations.js';
+} from '../../styles/foundations.js';
 import { actionRecorderDialogStyles } from './action-recorder-dialog.styles.js';
 
 const RECORDER_DIALOG_TAG = 'edvibe-toolbox-action-recorder';
