+++ src/content/main/features/execution-history/execution-history-dialog.js
@@ -2,7 +2,7 @@ import { LitElement, html } from 'lit';
 import {
     componentFoundationStyles,
     dialogFoundationStyles
-} from '../../../../components/styles/foundations.js';
+} from '../../styles/foundations.js';
 import { executionHistoryDialogStyles } from './execution-history-dialog.styles.js';
 
 const EXECUTION_HISTORY_DIALOG_TAG = 'edvibe-toolbox-execution-history-dialog';
