+++ src/content/main/styles/foundations.js
